feat: JPA 도메인 객체 정의
