+++ src/main/java/com/swp/board/domain/Comment.java
@@ -1,25 +1,24 @@
-package com.swp.common.domain;
-
+package com.swp.board.domain;
+import com.swp.user.domain.User;
 import org.springframework.data.annotation.CreatedDate;
-
 import javax.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "boards")
-public class Board {
+@Table(name = "comments")
+public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "board_id",nullable = false)
-    private Long boardId;
+    @Column(name = "comment_id",nullable = false)
+    private Long commentId;
+
+    @ManyToOne
+    @JoinColumn(name ="board_id",nullable = false)
+    private Board board;
 
     @ManyToOne
     @JoinColumn(name ="user_id",nullable = false)
-    private Long userId;
-
-    @ManyToOne
-    @JoinColumn(name ="study_id",nullable = false)
-    private Long studyId;
+    private User user;
 
     @Column(name="content",length=1000,nullable = true)
     private String content;
@@ -27,6 +26,4 @@ public class Board {
     @Column(name="created_at",nullable=false)
     @CreatedDate
     private LocalDateTime createdAt;
-
-
 }
