+++ src/main/java/com/swp/user/domain/RelationshipId.java
@@ -1,0 +1,7 @@
+package com.swp.user.domain;
+
+import java.io.Serializable;
+
+public class RelationshipId implements Serializable {
+    private Long fromUser,toUser;
+}
