+++ src/main/java/com/swp/user/domain/User.java
@@ -1,9 +1,10 @@
-package com.swp.common.domain;
-
+package com.swp.user.domain;
+import com.swp.board.domain.Board;
 import org.springframework.data.annotation.CreatedDate;
-
 import javax.persistence.*;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Table(name = "users")
@@ -26,4 +27,6 @@ public class User {
     @Column(name="deleted_at", nullable = true)
     private LocalDateTime deletedAt;
 
+    @OneToMany( mappedBy = "user")
+    private List<Board> boardList = new ArrayList<Board>();
 }
