+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+spring.jpa.show-sql=true
